Move utils tests too
